refactor(collector): split monolithic collector.py into collector/ package
Pure symbol migration (zero behavior change). The 1254-line collector.py
is decomposed into a package with a single-responsibility module per
concern:

  collector/
  ├── __init__.py     # compat shim: re-exports every legacy public/private
  │                   # symbol; preserves json/time/os module aliases on the
  │                   # package namespace so test monkeypatch surfaces work
  ├── __main__.py     # 'python -m collector' entrypoint
  ├── http.py         # requests.Session singleton + HEADERS + Retry policy
  ├── merge.py        # _better_str/_list/_code, _merge_paper_record,
  │                   # _merge_with_cache
  ├── progress.py     # COLLECT_PROGRESS_FILE, COLLECT_FAILURES_FILE,
  │                   # load/save_collect_progress
  ├── io.py           # cache.jsonl.gz load/save + _to_gz_path
  ├── code_links.py   # GitHub URL extraction from abstracts
  ├── pipeline.py     # collect() + do_collect() orchestration
  └── sources/
      ├── openreview.py  # OpenReview v1/v2 + ICLR schedule (search_from_iclr)
      ├── acl.py         # ACL Anthology volume pages
      ├── nips.py        # NeurIPS proceedings
      ├── dblp.py        # DBLP mixed venues
      └── thecvf.py      # CVF Open Access (deferred scripts.cvf_abstract import)

External consumers untouched (compat guaranteed by __init__.py re-exports):
  - maintain.py, papervault/services/papers.py
  - discovery/openreview.py
  - scripts/{cvf_abstract, cleanup_cache_dedupe, fetch_openreview_abstracts}.py

Verification: PAPERVAULT_OFFLINE=1 pytest -q -> 137 passed.
