+++ collector/__main__.py
@@ -1,0 +1,2 @@
+from collector import do_collect
+do_collect(cache_file="cache/cache.jsonl.gz", force=True)
