+++ collector/http.py
@@ -1,0 +1,29 @@
+import warnings
+
+import requests
+from requests.adapters import HTTPAdapter
+from urllib3.util.retry import Retry
+from bs4 import XMLParsedAsHTMLWarning
+
+# 忽略 ACL Anthology 某些 XML 页面被 HTML 解析器解析时的警告
+warnings.filterwarnings("ignore", category=XMLParsedAsHTMLWarning)
+
+HEADERS = {
+    "User-Agent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/100.0.4896.127 Safari/537.36"
+}
+
+
+def _create_session():
+    session = requests.Session()
+    retries = Retry(
+        total=5,
+        backoff_factor=1,
+        status_forcelist=[500, 502, 503, 504],
+        allowed_methods=["GET"],
+    )
+    session.mount("https://", HTTPAdapter(max_retries=retries))
+    session.mount("http://", HTTPAdapter(max_retries=retries))
+    return session
+
+
+SESSION = _create_session()
